+++ src/components/common/card/index.ts
@@ -1,0 +1,4 @@
+export { default as CardComponent } from "./default-card"
+export { default as TitleCard } from "./title-card"
+export { default as MiniCard } from "./mini-card"
+export { default as TitleContent } from "./content-card"
